client: modularized Step 2 component for ease of access
